requirements.txt.txt
flask
streamlit
mysql-connector-python
requests